+++ src/domain/NavItem.ts
