+++ src/presentation/components/atoms/Card.module.css
@@ -1,0 +1,6 @@
+.Card {
+  border-radius: 0.5rem;
+  padding: 1rem;
+  border: 1px solid #d0d7de;
+  width: 100%;
+}
