+++ src/presentation/components/atoms/Input.module.css
@@ -1,0 +1,6 @@
+.Input {
+  width: 100%;
+  padding: 0.2rem 0.8rem;
+  border-radius: 3px;
+  border: 1px solid #d0d7de;
+}
