+++ src/presentation/components/organisms/CreateJob/JobForm.tsx
@@ -2,9 +2,9 @@ import axios from "axios";
 import { ChangeEvent, FC, FormEvent, useEffect, useState } from "react";
 import { createJobUrl, fetchSkillsUrl } from "../../../../constants/constants";
 import { ISkill } from "../../../../domain/Skill";
-import CreateButton from "../../atoms/create-button";
-import Input from "../../atoms/input";
-import classes from "./job-form.module.css";
+import CreateButton from "../../atoms/CreateButton";
+import Input from "../../atoms/Input";
+import classes from "./JobForm.module.css";
 
 interface IProps {
   className: string;
