+++ src/presentation/components/organisms/Header.module.css
@@ -1,0 +1,22 @@
+.Header {
+  width: 100%;
+  padding: 0.5rem;
+  display: flex;
+  align-items: center;
+  border-bottom: 0.0625rem solid rgb(228, 226, 224);
+}
+
+.Header__logo-link {
+  text-decoration: none;
+  color: rgb(36, 87, 166);
+}
+
+.Header__logo-text {
+  font-size: 1.5rem;
+  margin-right: 1rem;
+}
+
+.Header__nav-items {
+  display: flex;
+  align-items: center;
+}
