+++ src/presentation/pages/JobList.tsx
@@ -1,9 +1,9 @@
 import axios from "axios";
 import { FC, useEffect, useState } from "react";
 import { Outlet } from "react-router-dom";
-import JobCardList from "../components/molecules/job-list/job-card-list";
+import JobCardList from "../components/molecules/JobList/JobCardList";
 import { IJob } from "../../domain/Job";
-import classes from "./job-list.module.css";
+import classes from "./JobList.module.css";
 
 export const jobsUrl = `${process.env.REACT_APP_BACKEND_URL}/jobs`;
